+++ wowPets/wow-pets-core/src/species/speciesService.js
@@ -1,8 +1,34 @@
 const speciesDao = require('./speciesDao');
+const {utils} = require('../other/utils');
+
+function checkSpecies(species){
+  if(Array.isArray(species)){
+    for (let i = 0; i < species.length; i++) {
+      const e = species[i];
+      if(e.tap && !utils.checkId(e.tap)){
+        return false;
+      }
+      if(e.hit && !utils.checkId(e.hit)){
+        return false;
+      }
+    }
+  }else {
+    if(species.tap && !utils.checkId(species.tap)){
+      return false;
+    }
+    if(species.hit && !utils.checkId(species.hit)){
+      return false;
+    }
+  }
+  return true;
+}
 
 const speciesService = {
   insert(species) {
     let result;
+    if(!checkSpecies(species)){
+      return utils.createResMsg(0,'id必须是24位字符串');
+    }
     if (Array.isArray(species)) {
       result = speciesDao.insertMany(species);
     } else {
@@ -11,7 +37,17 @@ const speciesService = {
     return result;
   },
   delete(_id) {
-    return speciesDao.deleteOne(_id);
+    let result,_ids=[];
+    _id.includes(',') ? _ids = _id.split(',') : _ids.push(_id);
+    if(!utils.checkId(_ids)){
+      return utils.createResMsg(0,'id必须是24位字符串');
+    }
+    if(_ids.length > 1){
+      result = speciesDao.deleteMany(_ids);
+    }else {
+      result = speciesDao.deleteOne(_id);
+    }
+    return result;
   },
   update(species) {
     return speciesDao.updateOne(species);
@@ -20,15 +56,17 @@ const speciesService = {
     return await speciesDao.find(where).then(species => {
       return species.map(entity => {
         species.map(prop => {
-          if(entity.tap == prop.code){
+          if(prop._id.equals(entity.tap)){
             entity.tapName = prop.name;
           }
-          if(entity.hit == prop.code){
+          if(prop._id.equals(entity.hit)){
             entity.hitName = prop.name;
           }
         });
         return entity;
       });
+    }).catch(error => {
+      return error;
     });
   }
 }
